-s modify

+++ ft_strlen.c
@@ -6,14 +6,14 @@
 /*   By: bosekim <bosekim@student.42seoul.k>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/13 21:42:39 by bosekim           #+#    #+#             */
-/*   Updated: 2023/03/13 22:12:50 by bosekim          ###   ########.fr       */
+/*   Updated: 2023/05/26 05:27:25 by bosekim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 size_t	ft_strlen(const char *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len])
